feat(109298): Remove warning E003
- Unifica logica do arquivo formata_valores do app de dre e core
